+++ example/lib/pdf/sample_pdf.dart
@@ -17,24 +17,27 @@ class SamplePdf {
             width: pageFormat.width * 0.7,
             height: pageFormat.height * 0.7,
             child: pw.Center(
-              child: pw.Text('''Sample pdf created by J_File_Saver Plugin
-              \n\n# rfilesaver
+              child: pw.Text('''# rfilesaver
 
-Plugin to save file in ANDROID or IOS device
+A Flutter Plugin to save file in ANDROID or IOS device
 
 ## Getting Started
 
 ## Minimal Sample
 
 ```dart
-String? path = await Rfilesaver.saveFile(
+SavedDetails savedDetails = await Rfilesaver.saveFile(
         data: data,
-        fileName: 'rfilesaver_sample',
+        fileName: "rfilesaver_sample",
         extension: MimeType.pdf.extension,
         mimeType: MimeType.pdf.mime,
+        pathToSave: pathToSave,
       );
 ```
-The MimeType has some predefined file types and its extension
+    The MimeType has some predefined file types and its extension
+    savedDetails holds 
+        String? savedPath,
+        String? contentUri,
 
 ## Android Setup
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"
@@ -46,7 +49,7 @@ The MimeType has some predefined file types and its extension
     * For android file will be stored in download folder
     * In Ios file will be stores in application documents directory
 
-## Used Dependencies
+## Dependencies Used
     permission_handler
     path_provider
     device_info_plus